+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added common command-line options to a ``parser`` module.
+- Added a ``clone_url`` property to ``InstalledProductVersion`` objects.
 
 ## [2.0.0] - 2021-10-06